chore: add WrappingHStack file
