+++ JWords/iOSView/items/EditableHuriganaText.swift
@@ -1,0 +1,20 @@
+//
+//  EditableHuriganaText.swift
+//  JWords
+//
+//  Created by JW Moon on 2023/04/30.
+//
+
+import SwiftUI
+
+struct EditableHuriganaText: View {
+    var body: some View {
+        Text(/*@START_MENU_TOKEN@*/"Hello, World!"/*@END_MENU_TOKEN@*/)
+    }
+}
+
+struct EditableHuriganaText_Previews: PreviewProvider {
+    static var previews: some View {
+        EditableHuriganaText()
+    }
+}
